@@ -25,9 +25,9 @@
     "payment": "^2.4.6",
     "prop-types": "^15.8.1",
     "qs": "^6.11.0",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-bootstrap": "^2.2.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-helmet-async": "^1.3.0",
     "react-hook-form": "^7.42.1",
     "react-phone-number-input": "^3.2.16",
